fix: Updated column value

+++ src/database/migrations/1593029149172-CreateTransactions.ts
@@ -25,7 +25,7 @@ export default class CreateTransactions1593029149172
           },
           {
             name: 'value',
-            type: 'decimal',
+            type: 'integer',
           },
           {
             name: 'type',
